src/requirements.txt
tzdata
requests~=2.28.1
Flask~=2.2.2
malclient-upgraded~=1.3.3
Jinja2~=3.1.2
pydantic~=1.10.2
Flask-Login~=0.6.2
celery
flask-celery
psycopg2-binary
flower
bs4
redis
gunicorn~=20.0
PyJWT
cryptography